requirements.txt
torch==2.1.0
torchvision==0.16.0
transformers==4.45.0
tokenizers==0.22.0
numpy<2.0
tiktoken
blobfile
clearml==2.0.2
ftfy==6.3.1
hydra-core==1.3.2
numpy==2.3.2
omegaconf==2.3.0
Pillow==11.3.0
pytorch_lightning==2.5.5
regex==2024.11.6
Requests==2.32.5
scikit_learn==1.7.1
setuptools==65.5.1
tensorboard==2.20.0
tqdm==4.67.1